Fix bug in RemainderMod4. Using a simpler approach.
