+++ openid-zkp-auth/circuits/misc.circom
@@ -3,7 +3,6 @@ pragma circom 2.0.0;
 include "../node_modules/circomlib/circuits/comparators.circom";
 include "../node_modules/circomlib/circuits/gates.circom";
 include "../node_modules/circomlib/circuits/mux2.circom";
-include "../node_modules/circomlib/circuits/poseidon.circom";
 
 // Returns math.ceil(log2(a)). Assumes a > 0.
 function log2(a) {
@@ -67,9 +66,6 @@ template IsEqualIfEnabled(n) {
     out <== and.out * enabled;
 }
 
-// This circuit returns the sum of the inputs.
-// n must be greater than 0.
-// Cost: No constraints.
 template CalculateTotal(n) {
     signal input nums[n];
     signal output sum;
@@ -84,9 +80,15 @@ template CalculateTotal(n) {
     sum <== sums[n - 1];
 }
 
-// TODO: Tests to be added.
 /**
-Packer: Packs a list of numbers into a list of numbers of a different size.
+Packer
+
+Packs a list of numbers, each of a specified bitwidth, into another list of numbers with a different bitwidth.
+
+- inWidth: The bitwidth of each input number.
+- inCount: The number of input numbers.
+- outWidth: The bitwidth of each output number.
+- outCount: The number of output numbers. (Should be inCount * inWidth / outWidth, rounded up.)
 */
 template Packer(inWidth, inCount, outWidth, outCount) {
     signal input in[inCount];
@@ -132,50 +134,26 @@ template Packer(inWidth, inCount, outWidth, outCount) {
     }
 }
 
-template Hasher(nInputs) {
-    signal input in[nInputs];
-    signal output out;
+/**
+RemainderMod4
 
-    component pos1, pos2, pos3;
-    if (nInputs <= 15) {
-        pos1 = Poseidon(nInputs);
-        for (var i = 0; i < nInputs; i++) {
-            pos1.inputs[i] <== in[i];
-        }
-        out <== pos1.out;
-    } else if (nInputs <= 30) {
-        pos1 = Poseidon(15);
-        pos2 = Poseidon(nInputs - 15);
+Calculates the remainder of a number divided by 4.
 
-        for (var i = 0; i < 15; i++) {
-            pos1.inputs[i] <== in[i];
-        }
-        for (var i = 15; i < nInputs; i++) {
-            pos2.inputs[i - 15] <== in[i];
-        }
+Construction Params:
+- maxBits: The maximum number of bits the input can have.
 
-        pos3 = Poseidon(2);
-        pos3.inputs[0] <== pos1.out;
-        pos3.inputs[1] <== pos2.out;
+IO Signals:
+- in: The input number.
+- out: The output remainder.
 
-        out <== pos3.out;
-    } else { // Yet to be implemented
-        1 === 0;
-    }
-}
-
-template RemainderMod4() {
+Assumptions:
+- maxBits < log(p) - 1, where p is the modulus of the field. Note that this automatically implies val(in) >= 0.
+**/
+template RemainderMod4(maxBits) {
     signal input in;
     signal output out;
 
-    out <-- in % 4;
-    signal q <-- in \ 4;
-    // TODO: Check if q is in range.
-
-    4 * q + out === in;
-
-    signal tmp1, tmp2;
-    tmp1 <== (out - 3) * (out - 2);
-    tmp2 <== tmp1 * (out - 1);
-    tmp2 * out === 0;
+    component toBits = Num2Bits(maxBits);
+    toBits.in <== in;
+    out <== 2 * toBits.out[1] + toBits.out[0];
 }
